Refactor pilot options handling
Add methods to read pilot options into the common space
and adapt firecracker and podman pilots to make use of
them
