+++ common/Cargo.toml
@@ -10,4 +10,5 @@ crate-type = ["lib"]
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
+log = { version = "0.4" }
 thiserror = { version = "1.0" }
